requirements.txt
boto3==1.38.13
pillow==11.2.1
selenium==4.32.0
